Lab_2/DL_Lab2_B11107122_凃岳霖/requirements.txt
# Python version requirement: >=3.9

# Deep Learning Core

torch>=2.0.0
torchvision>=0.15.0

# Data Manipulation & Image Processing

numpy>=1.24.0
pandas>=1.5.0
Pillow>=9.4.0

# Utilities

tqdm>=4.65.0
matplotlib>=3.7.0